Fix: Move tests
